Adding group feature for booking other users
Added feature to restrict the booking of other users to the same group
the user who does the booking belongs to.
In order to be able to do the booking the user have to have the
capability subscribeusers and group mode has to be set to forced. The user
must not have the capability moodle/site:accessallgroups. So standard role
for a user to book only members of his/her group would be non-editing teacher.