Settings to allow booking rules to send ics attachments (#801)
* Settings to allow booking rules to send ics attachments

Setting to toggle sending an ics attachment to booking_rules and a separate toggle to denoted if the attachments should an update or create ics.

Actual attachment is handled in the message controller. This should probably be moved to the method somewhere but I wasn't sure where.

* fix issues raised by CI

* We've noticed that ics files with html or new lines in them don't render correctly in Googlemail/ calendar, so I've added X-ALT-DESC to the ics.
I've also done a quick and dirty to render the urls as links.